// source/nodejs/adaptivecards/src/card-elements.d.ts
export interface IValidationError {
    error: Enums.ValidationError;
    message: string;
}
export declare abstract class CardElement {
    private _hostConfig?;
    private _internalPadding;
    private _parent;
    private _isVisibile;
    private _renderedElement;
    private _separatorElement;
    private internalRenderSeparator();
    private updateRenderedElementVisibility();
    protected internalGetNonZeroPadding(padding: HostConfig.PaddingDefinition): void;
    protected adjustRenderedElementSize(renderedElement: HTMLElement): void;
    protected showBottomSpacer(requestingElement: CardElement): void;
    protected hideBottomSpacer(requestingElement: CardElement): void;
    protected setParent(value: CardElement): void;
    protected readonly useDefaultSizing: boolean;
    protected abstract internalRender(): HTMLElement;
    protected readonly allowCustomPadding: boolean;
    protected readonly defaultPadding: HostConfig.PaddingDefinition;
    protected internalPadding: HostConfig.PaddingDefinition;
    protected readonly separatorOrientation: Enums.Orientation;
    id: string;
    speak: string;
    horizontalAlignment?: Enums.HorizontalAlignment;
    spacing: Enums.Spacing;
    separator: boolean;
    height: "auto" | "stretch";
    abstract getJsonTypeName(): string;
    abstract renderSpeech(): string;
    getNonZeroPadding(): HostConfig.PaddingDefinition;
    getForbiddenElementTypes(): Array<string>;
    getForbiddenActionTypes(): Array<any>;
    parse(json: any): void;
    validate(): Array<IValidationError>;
    render(): HTMLElement;
    updateLayout(processChildren?: boolean): void;
    isAtTheVeryTop(): boolean;
    isFirstElement(element: CardElement): boolean;
    isAtTheVeryBottom(): boolean;
    isLastElement(element: CardElement): boolean;
    isAtTheVeryLeft(): boolean;
    isLeftMostElement(element: CardElement): boolean;
    isAtTheVeryRight(): boolean;
    isRightMostElement(element: CardElement): boolean;
    canContentBleed(): boolean;
    getRootElement(): CardElement;
    getParentContainer(): Container;
    getAllInputs(): Array<Input>;
    getElementById(id: string): CardElement;
    getActionById(id: string): Action;
    hostConfig: HostConfig.HostConfig;
    readonly isInteractive: boolean;
    readonly isStandalone: boolean;
    readonly parent: CardElement;
    isVisible: boolean;
    readonly renderedElement: HTMLElement;
    readonly separatorElement: HTMLElement;
}
// ...
export declare abstract class Input extends CardElement implements Utils.IInput {
    id: string;
    title: string;
    defaultValue: string;
    readonly abstract value: string;
    validate(): Array<IValidationError>;
    parse(json: any): void;
    renderSpeech(): string;
    getAllInputs(): Array<Input>;
    readonly isInteractive: boolean;
}
// ...
export declare abstract class Action {
    private _parent;
    private _actionCollection;
    protected setParent(value: CardElement): void;
    abstract getJsonTypeName(): string;
    execute(): void;
    private setCollection(actionCollection);
    setStatus(status: any): void;
    validate(): Array<IValidationError>;
    prepare(inputs: Array<Input>): void;
    parse(json: any): void;
    getAllInputs(): Array<Input>;
    getActionById(id: string): Action;
    id: string;
    title: string;
    readonly parent: CardElement;
}
// ...
export declare class ShowCardAction extends Action {
    protected setParent(value: CardElement): void;
    readonly card: AdaptiveCard;
    title: string;
    getJsonTypeName(): string;
    validate(): Array<IValidationError>;
    parse(json: any): void;
    getAllInputs(): Array<Input>;
    getActionById(id: string): Action;
}
// ...
export declare class BackgroundImage {
    url: string;
    mode: Enums.BackgroundImageMode;
    horizontalAlignment: Enums.HorizontalAlignment;
    verticalAlignment: Enums.VerticalAlignment;
    parse(json: any): void;
    apply(element: HTMLElement): void;
}
export declare class Container extends CardElement {
    private _selectAction;
    private isElementAllowed(element, forbiddenElementTypes);
    private _items;
    private _style?;
    private readonly hasExplicitStyle;
    protected showBottomSpacer(requestingElement: CardElement): void;
    protected hideBottomSpacer(requestingElement: CardElement): void;
    protected applyPadding(): void;
    protected internalRender(): HTMLElement;
    protected readonly hasBackground: boolean;
    protected readonly defaultStyle: Enums.ContainerStyle;
    protected readonly allowCustomStyle: boolean;
    backgroundImage: BackgroundImage;
    bleed: boolean;
    verticalContentAlignment: Enums.VerticalAlignment;
    style: Enums.ContainerStyle;
    getJsonTypeName(): string;
    isFirstElement(element: CardElement): boolean;
    isLastElement(element: CardElement): boolean;
    validate(): Array<IValidationError>;
    parse(json: any, itemsCollectionPropertyName?: string): void;
    addItem(item: CardElement): void;
    clear(): void;
    canContentBleed(): boolean;
    getAllInputs(): Array<Input>;
    getElementById(id: string): CardElement;
    getActionById(id: string): Action;
    renderSpeech(): string;
    updateLayout(processChildren?: boolean): void;
    selectAction: Action;
}
// ...
export declare class Version {
    private _versionString;
    private _major;
    private _minor;
    private _isValid;
    constructor(major?: number, minor?: number);
    static parse(versionString: string): Version;
    toString(): string;
    readonly major: number;
    readonly minor: number;
    readonly isValid: boolean;
}
export declare abstract class ContainerWithActions extends Container {
    private _actionCollection;
    protected internalRender(): HTMLElement;
    constructor();
    getActionById(id: string): Action;
    parse(json: any, itemsCollectionPropertyName?: string): void;
    validate(): Array<IValidationError>;
    isLastElement(element: CardElement): boolean;
    addAction(action: Action): void;
    clear(): void;
    getAllInputs(): Array<Input>;
    readonly isStandalone: boolean;
}
export declare abstract class TypeRegistry<T> {
    private _items;
    private findTypeRegistration(typeName);
    constructor();
    clear(): void;
    abstract reset(): any;
    registerType(typeName: string, createInstance: () => T): void;
    unregisterType(typeName: string): void;
    createInstance(typeName: string): T;
}
export declare class ElementTypeRegistry extends TypeRegistry<CardElement> {
    reset(): void;
}
export declare class ActionTypeRegistry extends TypeRegistry<Action> {
    reset(): void;
}
export declare class AdaptiveCard extends ContainerWithActions implements IAdaptiveCard {
    private static currentVersion;
    static preExpandSingleShowCardAction: boolean;
    static readonly elementTypeRegistry: ElementTypeRegistry;
    static readonly actionTypeRegistry: ActionTypeRegistry;
    static onAnchorClicked: (anchor: HTMLAnchorElement) => boolean;
    static onExecuteAction: (action: Action) => void;
    static onElementVisibilityChanged: (element: CardElement) => void;
    static onInlineCardExpanded: (action: ShowCardAction, isExpanded: boolean) => void;
    static onParseElement: (element: CardElement, json: any) => void;
    static onParseError: (error: IValidationError) => void;
    private isVersionSupported();
    private _cardTypeName;
    protected applyPadding(): void;
    protected readonly bypassVersionCheck: boolean;
    protected readonly defaultPadding: HostConfig.PaddingDefinition;
    protected readonly allowCustomPadding: boolean;
    protected readonly allowCustomStyle: boolean;
    protected readonly hasBackground: boolean;
    version?: Version;
    fallbackText: string;
    type: string;
    getJsonTypeName(): string;
    validate(): Array<IValidationError>;
    parse(json: any): void;
    render(): HTMLElement;
    canContentBleed(): boolean;
}